+++ Example/Stopwatch/AppDelegate.m
@@ -54,7 +54,7 @@ static NSString *const AppboyApiKey = @"appboy-sample-ios";
     NSLog(@"Setting ABKInAppMessageControllerDelegate for app run.");
     appboyOptions[ABKInAppMessageControllerDelegateKey] = self;
   }
-  appboyOptions[ABKInAppMessageHideStatusBarKey] = @(YES);
+  appboyOptions[ABKInAppMessageHideStatusBarKey] = @YES;
 
   // Set ABKURLDelegate on startup
   BOOL setUrlDelegate = YES; // default value